Move the enabled count next to the tools that use it
All four call sites count tools: an MCP server's, an action's, the tools
list, and the agent editor's server card. Nothing outside that context uses
it, so it belongs in lib/tools/components rather than refresh-components.

Drops the noun prop. Only the tools list passed it, so three of the four
rendered a bare "3 of 12" while one read "3 of 12 tools". The surrounding
row already names what is counted, so the bare form is the one to keep and
the choice does not need to be a prop.

That leaves common.enabledCount.tools unused, so it goes from all nine
locales. The remaining label has no plural branch, since a bare count needs
no agreement.
