+++ web/src/i18n/messages/ar.json
@@ -10731,8 +10731,7 @@
       }
     },
     "enabledCount": {
-      "label": "<value>{enabled}</value> من {total}",
-      "tools": "{total, plural, one {<value>{enabled}</value> من {total} أداة} other {<value>{enabled}</value> من {total} أدوات}}"
+      "label": "<value>{enabled}</value> من {total}"
     },
     "errorPages": {
       "accessRestricted": {
