+++ web/src/i18n/messages/de.json
@@ -10731,8 +10731,7 @@
       }
     },
     "enabledCount": {
-      "label": "<value>{enabled}</value> von {total}",
-      "tools": "{total, plural, one {<value>{enabled}</value> von {total} Tool} other {<value>{enabled}</value> von {total} Tools}}"
+      "label": "<value>{enabled}</value> von {total}"
     },
     "errorPages": {
       "accessRestricted": {
