+++ web/src/i18n/messages/en.json
@@ -10731,8 +10731,7 @@
       }
     },
     "enabledCount": {
-      "label": "<value>{enabled}</value> of {total}",
-      "tools": "{total, plural, one {<value>{enabled}</value> of {total} tool} other {<value>{enabled}</value> of {total} tools}}"
+      "label": "<value>{enabled}</value> of {total}"
     },
     "errorPages": {
       "accessRestricted": {
