+++ web/src/i18n/messages/fr.json
@@ -10731,8 +10731,7 @@
       }
     },
     "enabledCount": {
-      "label": "<value>{enabled}</value> sur {total}",
-      "tools": "{total, plural, one {<value>{enabled}</value> sur {total} outil} other {<value>{enabled}</value> sur {total} outils}}"
+      "label": "<value>{enabled}</value> sur {total}"
     },
     "errorPages": {
       "accessRestricted": {
