+++ web/src/i18n/messages/ko.json
@@ -10731,8 +10731,7 @@
       }
     },
     "enabledCount": {
-      "label": "{total}개 중 <value>{enabled}</value>개",
-      "tools": "{total, plural, one {{total}개 도구 중 <value>{enabled}</value>개} other {{total}개 도구 중 <value>{enabled}</value>개}}"
+      "label": "{total}개 중 <value>{enabled}</value>개"
     },
     "errorPages": {
       "accessRestricted": {
