+++ web/src/i18n/messages/pt.json
@@ -10731,8 +10731,7 @@
       }
     },
     "enabledCount": {
-      "label": "<value>{enabled}</value> de {total}",
-      "tools": "{total, plural, one {<value>{enabled}</value> de {total} ferramenta} other {<value>{enabled}</value> de {total} ferramentas}}"
+      "label": "<value>{enabled}</value> de {total}"
     },
     "errorPages": {
       "accessRestricted": {
