+++ web/src/lib/tools/components/EnabledCount.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import EnabledCount from "./EnabledCount";
 
 const meta: Meta<typeof EnabledCount> = {
-  title: "refresh-components/EnabledCount",
+  title: "tools/EnabledCount",
   component: EnabledCount,
   tags: ["autodocs"],
   parameters: {
@@ -20,17 +20,8 @@ export const Default: Story = {
   },
 };
 
-export const WithNoun: Story = {
-  args: {
-    noun: "tool",
-    enabledCount: 3,
-    totalCount: 10,
-  },
-};
-
 export const AllEnabled: Story = {
   args: {
-    noun: "tool",
     enabledCount: 8,
     totalCount: 8,
   },
@@ -38,16 +29,13 @@ export const AllEnabled: Story = {
 
 export const NoneEnabled: Story = {
   args: {
-    noun: "tool",
     enabledCount: 0,
     totalCount: 15,
   },
 };
 
-/** The singular branch of the plural: "1 of 1 tool", not "tools". */
 export const SingleItem: Story = {
   args: {
-    noun: "tool",
     enabledCount: 1,
     totalCount: 1,
   },
