+++ web/src/lib/tools/components/ToolLineItem.tsx
@@ -29,7 +29,7 @@ import { useToolsPopover } from "@/lib/tools/providers";
 import { ToolSnapshot } from "@/lib/tools/types";
 import { getAdminConfigureInfo, getToolTooltip } from "@/lib/tools/utils";
 import { Permission } from "@/lib/types";
-import EnabledCount from "@/refresh-components/EnabledCount";
+import EnabledCount from "./EnabledCount";
 import { useUser } from "@/providers/UserProvider";
 
 // Names the hover group that swaps the source count for the disable button.
